Added support for ranges.
 120>            dl := dom{[ 1 .. Length(dom)-1 ]};
 120>                     ^

Next: add support for list composition.

list{ poss }

The above construct evaluates to a new list new whose first element is
list[poss[1]], whose second element is list[poss[2]], and so on.
However, it does not need to be sorted and may contain duplicate
elements. If for any i, list[ poss[i] ] is unbound, an error is
signalled.